Docker Package Version (called DockerImageVersion)
