+++ app/views/editStagePackagesJs.scala.html
@@ -15,7 +15,7 @@
 *@
 @(manifest: models.Manifest)
 
-@defining(manifest.getPackages.map((p: PackageVersion) => {s"""new Package("${p.getPkg.getName}", "${p.getVersion}")"""}).mkString(", ")) { initialPackages: String =>
+@defining(manifest.getPackages.map((p: RollerPackageVersion) => {s"""new Package("${p.getPkg.getName}", "${p.getVersion}")"""}).mkString(", ")) { initialPackages: String =>
 var initialPackages = [ @Html(initialPackages) ];
 }
